requirements.txt
asgiref==3.2.10
Django==3.1.2
djangorestframework==3.12.2
django-csp==3.7
pytz==2020.1
sqlparse==0.4.1
mozilla-django-oidc==1.2.4
django-configurations==2.2
